Fix description for colorMap
Function colorMap applies a function to all components.
Applying to colors would imply it doesn't apply the function
to the alpha plane of RGBA images, as it is questionable
whether transparency could be called a color.
It also has dubious meaning for Y, YCbCr etc.

It should probably be called componentMap, planeMap or something.