Parse extended bitmap headers
There are several extended versions of the BITMAPINFOHEADER.  We need to
parse these correctly so that we know where the palette starts and where
the bitfields are.

The extended headers contain color profiles which I have opted to
expose as metadata.  Other image formats (e.g. PNG and JPEG) support
similar color profiles and it would be a good idea to expose the color
profile information from those formats also.